product search is a bit more capable now
- item-search template has improved search box that accepts submission
- item-search component navigates to '/products/;term' on submission
- items component aware of route param 'term' and if param['term'] has
been set, uses that to set this.filter. Otherwise in the case where
param['term'] hasn't been set, it's assumed that property binding
triggered the component and this.filter has already been set. Finally
this.filter is used to properly set the items
- app-routing module now has 'products/:term' and redirects 'products'
route to 'products/:term' where ':term' is set to 'all'